yoloe-box-prompting/pyproject.toml
[project]
name = "yoloe-box-prompting"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ultralytics>=8.3.202",
    "pillow>=10.0.0"
]
